movimiento 2 fichas paso a paso, compra de propiedades
disminución de saldo, asignación del turno para cada jugador,
disminución tamaño imágenes, cambio de color de fondo segun jugador

+++ Proyecto Monopoly/MonopolyCR/MonopolyCR/Jugador.cs
@@ -21,7 +21,7 @@ namespace MonopolyCR
 		public int Puntuacion  { get; set; }
 		public String Color { get; set; }
         public int PosicionActual { get; set; }
-        public int Saldo { get; set; }
+        public double Saldo { get; set; }
 
 	
 		public Jugador()
